// src/renderer/store/types/PadPrm.ts
export interface PadPrmType {
    Wv: number;
    WvLevel: number;
    WvPan: number;
    PlayMode: number;
    OutAsgn: number;
    MuteGrp: number;
    TempoSync: number;
    PadMidiCh: number;
    NoteNum: number;
    MidiCtrl: number;
    Loop: number;
    TrigType: number;
    GateTime: number;
    Dynamics: number;
    VoiceAsgn: number;
    Reverse: number;
    SubWv: number;
    SubWvLevel: number;
    SubWvPan: number;
}